+++ src/Accounts/Accounts/Registration/RegistrationController.cs
@@ -22,9 +22,7 @@
  */
 
 using System;
-using System.Collections.Generic;
 using System.Text;
-using MessagePack.Internal;
 using Sodium;
 using Sovereign.Persistence.Database;
 using Sovereign.Persistence.Database.Queries;
@@ -37,11 +35,31 @@ namespace Sovereign.Accounts.Accounts.Registration
     /// </summary>
     public sealed class RegistrationController
     {
-        private readonly IAddAccountQuery query;
+
+        /// <summary>
+        /// Argon2 opslimit.
+        /// </summary>
+        /// <remarks>
+        /// Ideally the "Interactive" constant would be used,
+        /// but the design of the Sodium.Core bindings unfortunately precludes
+        /// this - it is possible to use these constants, but not to retrieve
+        /// the values and store them alongside the hash in the database.
+        /// </remarks>
+        private const long OPSLIMIT = 4;
+
+        /// <summary>
+        /// Argon2 memlimit.
+        /// </summary>
+        /// <remarks>
+        /// See remarks on OPSLIMIT.
+        /// </remarks>
+        private const int MEMLIMIT = 33554432;
+
+        private readonly PersistenceProviderManager providerManager;
 
         public RegistrationController(PersistenceProviderManager providerManager)
         {
-            query = providerManager.PersistenceProvider.AddAccountQuery;
+            this.providerManager = providerManager;
         }
 
         /// <summary>
@@ -58,22 +76,35 @@ namespace Sovereign.Accounts.Accounts.Registration
                 out var opsLimit, out var memLimit);
 
             /* Create account. */
+            var query = providerManager.PersistenceProvider.AddAccountQuery;
             return query.AddAccount(Guid.NewGuid(), username,
                 salt, hash, opsLimit, memLimit);
         }
 
+        /// <summary>
+        /// Generates a salt for a password hash.
+        /// </summary>
+        /// <returns>Salt.</returns>
         private byte[] GenerateSalt()
         {
             return PasswordHash.ArgonGenerateSalt();
         }
 
+        /// <summary>
+        /// Hashes a password.
+        /// </summary>
+        /// <param name="password">Password to be hashed.</param>
+        /// <param name="salt">Password salt.</param>
+        /// <param name="hash">Password hash.</param>
+        /// <param name="opsLimit">Opslimit parameter.</param>
+        /// <param name="memLimit">Memlimit parameter.</param>
         private void HashPassword(string password, byte[] salt, out byte[] hash,
             out ulong opsLimit, out ulong memLimit)
         {
-            opsLimit = (ulong)PasswordHash.StrengthArgon.Interactive;
-            memLimit = (ulong) PasswordHash.StrengthArgon.Interactive;
+            opsLimit = (ulong)OPSLIMIT;
+            memLimit = (ulong)MEMLIMIT;
             hash = PasswordHash.ArgonHashBinary(Encoding.UTF8.GetBytes(password), 
-                salt, (long)opsLimit, (int)memLimit);
+                salt, OPSLIMIT, MEMLIMIT);
         }
 
     }